requirements.txt
streamlit==1.38.0
pandas==2.2.3 
requests==2.32.3 
azure-identity==1.18.0 
openai
plotly
python-dotenv==1.0.1 
pytz==2024.2 
matplotlib==3.9.2
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0